refactor: simplify secure iframe architecture and enhance functionality
- Enhanced secure-iframe.ts with postMessage handling and progress tracking
- Updated bookmark-reader.ts to use secure-iframe directly
- Removed secure-iframe-container.ts to simplify component hierarchy
- Updated integration tests to work with new architecture
- Cleaned up unit tests for better maintainability
