+++ src/test/unit/security-service.test.ts
@@ -207,57 +207,4 @@ describe('SecurityService', () => {
     });
   });
 
-  describe('validateContent', () => {
-    it('should return true for valid HTML content', () => {
-      const validHtml = `
-        <!DOCTYPE html>
-        <html>
-        <head><title>Test</title></head>
-        <body><h1>Content</h1></body>
-        </html>
-      `;
-      
-      expect(SecurityService.validateContent(validHtml)).toBe(true);
-    });
-
-    it('should return true for content with body tag', () => {
-      const htmlWithBody = '<body><h1>Content</h1></body>';
-      
-      expect(SecurityService.validateContent(htmlWithBody)).toBe(true);
-    });
-
-    it('should return false for empty content', () => {
-      expect(SecurityService.validateContent('')).toBe(false);
-    });
-
-    it('should return false for null content', () => {
-      expect(SecurityService.validateContent(null as any)).toBe(false);
-    });
-
-    it('should return false for non-HTML content', () => {
-      expect(SecurityService.validateContent('Plain text content')).toBe(false);
-    });
-  });
-
-  describe('createNetworkTestPage', () => {
-    it('should create a test page with network blocking tests', () => {
-      const testPage = SecurityService.createNetworkTestPage();
-      
-      expect(testPage).toContain('<!DOCTYPE html>');
-      expect(testPage).toContain('Network Blocking Test');
-      expect(testPage).toContain('fetch(');
-      expect(testPage).toContain('XMLHttpRequest');
-      expect(testPage).toContain('WebSocket');
-      expect(testPage).toContain('https://example.com');
-    });
-
-    it('should include test result indicators', () => {
-      const testPage = SecurityService.createNetworkTestPage();
-      
-      expect(testPage).toContain('✅');
-      expect(testPage).toContain('❌');
-      expect(testPage).toContain('blocked');
-      expect(testPage).toContain('not blocked');
-    });
-  });
 });
